Make MediaConnector extensible.
